move Constant class
