+++ cfdm/__init__.py
@@ -78,7 +78,7 @@ if LooseVersion(netcdf_flattener.__version__) < LooseVersion(_minimum_vn):
             _minimum_vn,
             netcdf_flattener.__version__, netcdf_flattener.__file__))
 
-from .constants import Constant, masked
+from .constants import masked
 
 # Internal ones passed on so they can be used in cf-python (see
 # comment below)
@@ -98,6 +98,7 @@ from .functions import (
     _reset_log_emergence_level,
     _is_valid_log_level_int,
     Configuration,
+    Constant,
 )
 
 # Though these are internal-use methods, include them in the namespace
