+++ cfdm/functions.py
@@ -4,6 +4,8 @@ import platform
 import sys
 import urllib.parse
 
+from functools import total_ordering
+
 import netCDF4
 import cftime
 import numpy
@@ -13,7 +15,7 @@ from . import (__version__,
                __cf_version__,
                __file__)
 
-from .constants import CONSTANTS, ValidLogLevels, Constant
+from .constants import CONSTANTS, ValidLogLevels
 
 
 def configuration(atol=None, rtol=None, log_level=None):
@@ -589,7 +591,7 @@ def environment(display=True, paths=True):
 
     **Examples:**
 
-    >>> environment()
+    >>> cfdm.environment()
     Platform: Linux-4.15.0-72-generic-x86_64-with-debian-stretch-sid
     HDF5 library: 1.10.2
     netcdf library: 4.6.1
@@ -599,7 +601,7 @@ def environment(display=True, paths=True):
     numpy: 1.16.2 /home/user/anaconda3/lib/python3.7/site-packages/numpy/__init__.py
     cfdm: 1.8.8.0
 
-    >>> environment(paths=False)
+    >>> cfdm.environment(paths=False)
     Platform: Linux-4.15.0-72-generic-x86_64-with-debian-stretch-sid
     HDF5 library: 1.10.2
     netcdf library: 4.6.1
@@ -666,7 +668,7 @@ def CF():
 
     **Examples:**
 
-    >>> CF()
+    >>> cfdm.CF()
     '1.8'
 
     '''
@@ -696,11 +698,11 @@ def abspath(filename):
     >>> import os
     >>> os.getcwd()
     '/data/archive'
-    >>> abspath('file.nc')
+    >>> cfdm.abspath('file.nc')
     '/data/archive/file.nc'
-    >>> abspath('..//archive///file.nc')
+    >>> cfdm.abspath('..//archive///file.nc')
     '/data/archive/file.nc'
-    >>> abspath('http://data/archive/file.nc')
+    >>> cfdm.abspath('http://data/archive/file.nc')
     'http://data/archive/file.nc'
 
     '''
@@ -709,6 +711,304 @@ def abspath(filename):
         return filename
 
     return os.path.abspath(filename)
+
+
+@total_ordering
+class Constant:
+    '''A container for a constant with context manager support.
+
+    The constant value is accessed via the `value` attribute:
+
+       >>> c = cfdm.Constant(1.9)
+       >>> c.value
+       1.9
+
+    Conversion to `int`, `float` and `str` is with the usual built-in
+    functions:
+
+       >>> c = cfdm.Constant(1.9)
+       >>> int(c)
+       1
+       >>> float(c)
+       1.9
+       >>> str(c)
+       '1.9'
+
+    Augmented arithmetic assignments (``+=``, ``-=``, ``*=``, ``/=``,
+    ``//=``) update `Constant` objects in-place:
+
+       >>> c = cfdm.Constant(20)
+       >>> c.value
+       20
+       >>> c /= 2
+       >>> c
+       <Constant: 10.0>
+       >>> c += c
+       <Constant: 20.0>
+
+       >>> c = cfdm.Constant('New_')
+       >>> c *= 2
+       <Constant: 'New_New_'>
+
+    Binary arithmetic operations (``+``, ``-``, ``*``, ``/``, ``//``)
+    are equivalent to the operation acting on the `Constant` object's
+    `value` attribute:
+
+       >>> c = cfdm.Constant(20)
+       >>> c.value
+       20
+       >>> c * c
+       400
+       >>> c * 3
+       60
+       >>> 2 - c
+       -38
+
+       >>> c = cfdm.Constant('New_')
+       >>> c * 2
+       'New_New_'
+
+    Care is required when the right hand side operand is a `numpy`
+    array
+
+       >>> import numpy
+       >>> c * numpy.array([1, 2, 3])
+       array([20, 40, 60])
+       >>> d = numpy.array([1, 2, 3]) * c
+       >>> d
+       array([10, 20, 30], dtype=object)
+       >>> type(d[0])
+       int
+
+    Unary arithmetic operations (``+``, ``-``, `abs`) are equivalent
+    to the operation acting on the `Constant` object's `value`
+    attribute:
+
+       >>> c = cfdm.Constant(-20)
+       >>> c.value
+       -20
+       >>> -c
+       20
+       >>> abs(c)
+       20
+       >>> +c
+       -20
+
+    Rich comparison operations are equivalent to the operation acting
+    on the `Constant` object's `value` attribute:
+
+       >>> c = cfdm.Constant(20)
+       >>> d = cfdm.Constant(1)
+       >>> c.value
+       20
+       >>> d.value
+       1
+       >>> d < c
+       True
+       >>> c != 20
+       False
+       >>> 20 == c
+       True
+
+       >>> c = cfdm.Constant('new')
+       >>> d = cfdm.Constant('old')
+       >>> c == d
+       False
+       >>> c == 'new'
+       True
+       >>> c != 3
+       True
+       >>> 3 == c
+       False
+
+       >>> import numpy
+       >>> c = cfdm.Constant(20)
+       >>> c < numpy.array([10, 20, 30])
+       array([False, False,  True])
+       >>> numpy.array([10, 20, 30]) >= c
+       array([False,  True,  True])
+
+    `Constant` instances are hashable.
+
+    **Context manager**
+
+    The `Constant` instance can be used as a context manager that upon
+    exit executes the function defined by the `_func` attribute, with
+    the `value` attribute as an argument. For example, the `Constant`
+    instance ``c`` would execute ``c._func(c.value)`` upon exit.
+
+    .. versionadded:: (cfdm) 1.8.8.0
+
+    '''
+    __slots__ = ('_func',  'value', '_type')
+
+    def __init__(self, value, _func=None):
+        '''**Initialization**
+
+    :Parameters:
+
+        value:
+            A value for the constant.
+
+        _func: function, optional
+            A function that that is executed upon exit from a context
+            manager, that takes the *value* parameter as its argument.
+
+        '''
+        self.value = value
+        self._func = _func
+
+    def __enter__(self):
+        '''Enter the runtime context.
+
+        '''
+        if getattr(self, '_func', None) is None:
+            raise AttributeError(
+                "Can't use {!r} as a context manager because the '_func' "
+                "attribute is not defined".format(self))
+
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        '''Exit the runtime context.
+
+        '''
+        self._func(self.value)
+
+    def __deepcopy__(self, memo):
+        '''Called by the `copy.deepcopy` function.
+
+    x.__deepcopy__() <==> copy.deepcopy(x)
+
+        '''
+        return self.copy()
+
+    def __float__(self):
+        return float(self.value)
+
+    def __int__(self):
+        return int(self.value)
+
+    def __eq__(self, other):
+        return self.value == other
+
+    def __lt__(self, other):
+        return self.value < other
+
+    def __abs__(self):
+        return abs(self.value)
+
+    def __neg__(self):
+        return -self.value
+
+    def __pos__(self):
+        return self.value
+
+    def __iadd__(self, other):
+        self.value += other
+        return self
+
+    def __ifloordiv__(self, other):
+        self.value //= other
+        return self
+
+    def __imul__(self, other):
+        self.value *= other
+        return self
+
+    def __isub__(self, other):
+        self.value -= other
+        return self
+
+    def __itruediv__(self, other):
+        self.value /= other
+        return self
+
+    def __add__(self, other):
+        try:
+            return self.value + other
+        except TypeError:
+            return NotImplemented
+
+    def __floordiv__(self, other):
+        try:
+            return self.value // other
+        except TypeError:
+            return NotImplemented
+
+    def __mul__(self, other):
+        try:
+            return self.value * other
+        except TypeError:
+            return NotImplemented
+
+    def __sub__(self, other):
+        try:
+            return self.value - other
+        except TypeError:
+            return NotImplemented
+
+    def __truediv__(self, other):
+        try:
+            return self.value / other
+        except TypeError:
+            return NotImplemented
+
+    def __radd__(self, other):
+        return other + self.value
+
+    def __rfloordiv__(self, other):
+        return other // self.value
+
+    def __rmul__(self, other):
+        return other * self.value
+
+    def __rsub__(self, other):
+        return other - self.value
+
+    def __rtruediv__(self, other):
+        return other / self.value
+
+    def __hash__(self):
+        return hash((self.value, getattr(self, '_func', None)))
+
+    def __repr__(self):
+        '''Called by the `repr` built-in function.
+
+        '''
+        return "<{0}: {1!r}>".format(
+            self.__class__.__name__, self.value
+        )
+
+    def __str__(self):
+        '''Called by the `str` built-in function.
+
+        '''
+        return str(self.value)
+
+    # ----------------------------------------------------------------
+    # Methods
+    # ----------------------------------------------------------------
+    def copy(self):
+        '''Return a deep copy.
+
+    ``c.copy()`` is equivalent to ``copy.deepcopy(c)``.
+
+    .. versionadded:: (cfdm) 1.8.8.0
+
+    :Returns:
+
+        `Constant`
+            The deep copy.
+
+        '''
+        out = type(self)(value=deepcopy(self.value),
+                         _func=getattr(self, '_func', None))
+
+        if not hasattr(self, '_func'):
+            del out._func
+
+        return out
 
 
 class Configuration(dict):
